# docs/openapi/phase1-local.yaml
openapi: 3.1.0
info:
  title: IDA local Phase 1 API
  version: 0.1.0
  description: >-
    Contrat du runtime de développement local. Il expose uniquement des données
    de démonstration et des commandes IDA déterministes en lecture seule.
servers:
  - url: http://127.0.0.1:8787
    description: API locale de développement
paths:
...
  /v1/artist-profile:
    get:
      operationId: getArtistProfile
      summary: Retourne le profil artistique de démonstration
      responses:
        '200':
          description: Artist Brain de démonstration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistProfileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/media:
    get:
      operationId: listMedia
      summary: Liste les médias de démonstration
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MediaStatus'
      responses:
        '200':
          description: Médias filtrés dans le workspace local
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaListResponse'
        '400':
          $ref: '#/components/responses/InvalidCommand'
...
  /v1/ida/commands:
    post:
      operationId: executeIdaCommand
      summary: Exécute une commande IDA déterministe et en lecture seule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdaCommandRequest'
      responses:
        '200':
          description: Commande comprise et exécutée avec des outils READ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdaCommandResponse'
        '400':
          $ref: '#/components/responses/InvalidCommand'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  responses:
    InvalidCommand:
      description: Corps de commande ou filtre invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource de démonstration introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Outil non autorisé par la policy serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    ArtistProfile:
      type: object
      required: [id, workspaceId, artistProjectId, identity, genres, influences, tone, goals, audience, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        artistProjectId:
          type: string
        identity:
          type: string
        genres:
          type: array
          items: { type: string }
        influences:
          type: array
          items: { type: string }
        tone:
          type: string
        goals:
          type: array
          items: { type: string }
        audience:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArtistProfileResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ArtistProfile'
...
    MediaStatus:
      enum: [UNUSED, USED, SCHEDULED, PUBLISHED, ARCHIVED]
    MediaAsset:
      type: object
      required: [id, workspaceId, artistProjectId, filename, type, mimeType, size, hash, tags, status, usageCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        artistProjectId: { type: string }
        filename: { type: string }
        type:
          enum: [AUDIO, IMAGE, VIDEO, DOCUMENT, OTHER]
        mimeType: { type: string }
        size: { type: integer, minimum: 0 }
        hash: { type: string }
        tags:
          type: array
          items: { type: string }
        status:
          $ref: '#/components/schemas/MediaStatus'
        usageCount: { type: integer, minimum: 0 }
        lastUsedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MediaListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MediaAsset'
...
    IdaCommandRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
      additionalProperties: true
    IdaCommandResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [command, kind, message, tools, result]
          properties:
            command:
              type: object
              required: [id, workspaceId, message, intent, state, requestedPermission, createdAt, updatedAt]
              properties:
                id: { type: string }
                workspaceId: { type: string }
                message: { type: string }
                intent:
                  enum: [PREPARE_DAY, LIST_UNUSED_CONTENT, UNKNOWN]
                state:
                  const: COMPLETED
                requestedPermission:
                  const: READ
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
            kind:
              enum: [TODAY, UNUSED_CONTENT, SYSTEM, HELP]
            message: { type: string }
            tools:
              type: array
              items:
                type: object
                required: [key, moduleKey, permission]
                properties:
                  key: { type: string }
                  moduleKey:
                    enum: [TASKS, CONTENT, SYSTEM, IDA]
                  permission:
                    const: READ
            result:
              type: object
              additionalProperties: true